Split ids tests
